Code reorganization
Addressing CI comments and code reorganization for easier understanding.
